Minor error handling fixes
When a server is in active accept mode, errors are silently discarded. This is not necessarily a problem, but when a TConnection implementation did not have a valid constructor or an erroring constructor the connection would be accepted but silently fail. This resulted in a floating TCP client that would not close until the client went anyway.

- Invalid constructors would cause a connection to be accepted but not trigger any other logic, making it seem like the server had frozen. The server now validates that TConnection has a valid constructor
- Constructors which threw exceptions were also a problem, as if most constructions succeeded the failed executions would result in unclosable floating TCP clients. The accept function now closes the client if the connection object cannot be instantiated